# static/swagger.yaml
openapi: 3.0.3
info:
  title: md-to-pdf Document Engine API
  description: |
    Moteur de documents complet : conversion Markdown/HTML/Template vers PDF,
    thèmes, graphiques et diagrammes, aperçu PNG, fusion, filigrane, protection
    par mot de passe, caviardage, audit de mise en page et comparaison visuelle.

    ## Comportement conditionnel des réponses

    La plupart des endpoints de génération PDF acceptent les champs optionnels
    `client_id` et `pdf_name`. Selon leur présence :

    | `client_id` + `pdf_name` | Réponse |
    |--------------------------|---------|
    | **fournis** | Le PDF est sauvegardé sur le serveur et un JSON `{ "download_url": "/download/…" }` est retourné |
    | **absents** | Le fichier PDF binaire est retourné directement dans le body |

    `client_id` et `pdf_name` servent à construire un chemin sur disque : ils doivent être
    des noms simples (`[A-Za-z0-9._-]`, 128 caractères max, ne commençant pas par un point).

    ## Deux façons de désigner un PDF

    Partout où un corps de requête attend un PDF — `pdf`, `before`, `after`, les entrées de
    `pdfs` — **deux formes sont acceptées** :

    | Référence | Origine |
    |-----------|---------|
    | `/download/{client_id}/{nom}.pdf` | un PDF que ce service a produit et sauvegardé |
    | `asset://as_…` | un fichier déposé par l'appelant via `POST /api/files` |

    La première n'a pas changé et ne changera pas. La seconde est ce qui rend les outils
...

    ## Sortie des outils

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - ApiKeyHeader: []
  - BearerToken: []
...
paths:
...
  /api/protect:
    post:
      tags: [Protect]
      summary: Protéger un PDF par mot de passe
      description: |
        Chiffre un PDF existant avec un mot de passe (AES 256 bits via qpdf).

        Le PDF source est un chemin `/download/…` **ou** une référence `asset://as_…`.
      operationId: protect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProtectRequest"
            example:
              pdf: "/download/client-123/rapport.pdf"
              password: "secret123"
              client_id: "client-123"
              pdf_name: "rapport-protege"
      responses:
        "200":
          description: PDF protégé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConvertResponse"
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "504":
          $ref: "#/components/responses/Timeout"

  # ───────────────────────────────────────────────
  # Redact
  # ───────────────────────────────────────────────
...
components:
...
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Clé d'API, active uniquement si `API_KEY` est définie côté serveur.
    BearerToken:
      type: http
      scheme: bearer
      description: |
        Même clé que `X-API-Key`, transmise via `Authorization: Bearer <clé>`.
    SessionCookie:
      type: apiKey
      in: cookie
      name: mdpdf_session
      description: |
        Cookie de session posé par `POST /api/auth/signup` et `POST /api/auth/login`.
        `HttpOnly`, `SameSite=Lax`, et `Secure` dès que `X-Forwarded-Proto` vaut `https`.

        Il ne se mélange jamais avec `X-API-Key` : le cookie dit « un humain regarde une
        page », l'en-tête dit « un programme appelle ». C'est cette séparation qui permet au
        cookie de rester `Lax` — donc de refuser de voyager sur un POST cross-site — pendant
        que la clé reste utilisable depuis n'importe où.

  # ─── Paramètres réutilisables ─────────────────
...
  schemas:
    PdfSource:
      type: string
      description: |
        Référence d'un PDF, sous **l'une des deux formes acceptées** :

        - `/download/{client_id}/{nom}.pdf` — un PDF que ce service a produit et sauvegardé ;
        - `asset://as_…` — un fichier déposé via `POST /api/files`.

        La première forme est celle d'hier, inchangée. La seconde est ce qui rend l'endpoint
        utilisable sur un document que le service n'a pas fabriqué.
      example: "asset://as_9f2c4b1d8e6a03f75c21b40d9e8a1c37"
...
    ProtectRequest:
      type: object
      required: [pdf, password]
      properties:
        pdf:
          $ref: "#/components/schemas/PdfSource"
        password:
          type: string
          description: Mot de passe pour le chiffrement AES-256
        client_id:
          type: string
        pdf_name:
          type: string
...
    ConvertResponse:
      type: object
      description: |
        Strictement additive : une requête qui ne demande rien de nouveau reçoit
        exactement `{ "download_url": "…" }`, comme avant.
      properties:
        download_url:
          type: string
          description: URL relative pour télécharger le PDF généré
          example: "/download/client-123/rapport-2024.pdf"
        cached:
          type: boolean
          description: Présent (et vrai) uniquement quand le PDF vient du cache de rendu
        layout:
          allOf:
            - $ref: "#/components/schemas/LayoutReport"
          description: Présent uniquement quand `options.autolayout` a été demandé
        warnings:
          type: array
          description: |
            Présent uniquement s'il y en a : un bloc `chart` ou `mermaid` qui n'a pas pu
            être rendu est laissé tel quel dans le document et la raison arrive ici.
          items:
            $ref: "#/components/schemas/BlockWarning"
    BlockWarning:
      type: object
      properties:
        kind:
          type: string
          enum: [chart, mermaid]
        message:
          type: string
        line:
          type: integer
          description: Ligne 1-indexée du bloc fautif dans le document source
    LayoutReport:
      type: object
      properties:
        pages:
          type: integer
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: 100 = aucun défaut détecté
        passes:
          type: integer
          description: Nombre de passes correctives appliquées (autolayout uniquement)
        issues:
          type: array
          items:
            type: object
            properties:
              kind:
                type: string
                enum: [overflow, blank_page, widow_page, orphan_heading, split_table, long_table]
              page:
                type: integer
                description: Numéro de page 1-indexé
              severity:
                type: string
                enum: [error, warn, info]
              detail:
                type: string
              bbox:
                type: array
                description: x0, y0, x1, y1 en points PDF, quand le défaut est localisé
                items:
                  type: number

    # ─── Assets et travaux ──────────────────────
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Type d'erreur
          example: "Bad request"
        details:
          type: string
          description: Détails / message d'erreur
          example: "At least 2 PDFs are required for merging"
...
  responses:
...
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Bad request"
            details: "\"data\" must be a JSON object"
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Not found"
            details: "PDF not found: /download/client-123/missing.pdf"
    Unauthorized:
      description: Clé d'API manquante ou invalide (uniquement si `API_KEY` est configurée)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Unauthorized"
            details: "Missing or invalid API key (X-API-Key or Authorization: Bearer)"
    TooManyRequests:
      description: |
        La file de rendu est pleine : la requête a attendu plus de
        `PDF_QUEUE_TIMEOUT_SECS` sans obtenir de place. Réessayer quelques secondes plus tard.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Timeout:
      description: La génération a dépassé la limite de temps (`PDF_PROCESS_TIMEOUT_SECS`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Timeout"
            details: "weasyprint exceeded the 60s time limit"
    InternalError:
      description: Erreur interne du serveur
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Pandoc conversion failed"
            details: "pandoc: ... (stderr output)"